Fix async xhr issues in safari
From what I understand, the `loadend` event for an async xhr request
will fire once progress has stopped, not when the request is fully
ready. So trying to access the response text at this point causes
safari to throw an error. It doesn't cause any issues on the page,
just pollutes the console with error messages.